+++ scenes/board/logic/scripting/outcomes/message.gd
@@ -4,6 +4,7 @@ var text
 var portrait = null
 var name
 var side = "left"
+var colour = null
 
 func _execute(_metadata):
     var actor = {
@@ -15,6 +16,13 @@ func _execute(_metadata):
 
     actor['portrait_tile'].tile_view_height_cam_modifier = -0.2
 
+    if self.colour != null:
+        var material_type = self.board.map.templates.MATERIAL_NORMAL
+        if actor['portrait_tile'].uses_metallic_material:
+            material_type = self.board.map.templates.MATERIAL_METALLIC
+
+        actor['portrait_tile'].set_side_materials(self.board.map.templates.get_side_material(self.colour, material_type), self.board.map.templates.get_side_material_desat(self.colour, material_type))
+
     self.board.ui.show_story_dialog(text, actor)
 
 func _ingest_details(details):
@@ -24,3 +32,5 @@ func _ingest_details(details):
     self.name = details['name']
     if details.has("side"):
         self.side = details['side']
+    if details.has("colour"):
+        self.colour = details['colour']