+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### 0.4.0-alpha
 - Added Amber Noon campaign
+- Added new tiles
+- Improvements to story triggers
 
 ### 0.3.1-alpha
 - Cleaned-up Settings window